refactor : main controller 에서 input에 대한 책임을 하위 input controller에게 위임.
