+++ src/main/java/racing/controller/input/InputView.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.controller.input;
 
 import java.util.Scanner;
